requirements.txt
streamlit>=1.31.0
chromadb>=0.4.22
sentence-transformers>=2.3.1
httpx>=0.26.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0 